refactor(phase3c): accelerator setup を library/accelerator_setup.py に分離
train_util.py から以下を library/accelerator_setup.py に移動:
- HIGH_VRAM (mutable module-level flag)
- enable_high_vram
- prepare_dataset_args
- prepare_accelerator
- prepare_dtype
- patch_accelerator_for_fp16_training

train_util.py は関数の re-export shim を維持しつつ、HIGH_VRAM 属性アクセスは
module-level __getattr__ で accelerator_setup.HIGH_VRAM へ動的転送する。
これにより train_util.HIGH_VRAM 形式の参照（外部スクリプト互換性）でも
最新値が読み取れる。

リポジトリ内コードは shim に依存しない方針に従い、HIGH_VRAM の参照を
新パス直接参照に切替:
- library/args.py: enable_high_vram の参照
- library/caching.py: _tu().HIGH_VRAM → accelerator_setup.HIGH_VRAM
  (HIGH_VRAM のためだけに残っていた _tu() ヘルパーは削除)
- library/dataset.py: 同上 (_tu() ヘルパー削除)
- library/strategy_{sd,flux,sd3,lumina,anima,hunyuan_image}.py 6 ファイル:
  train_util.HIGH_VRAM → accelerator_setup.HIGH_VRAM

train_util.py: 2,103 → 1,969 行
library/accelerator_setup.py: 186 行（新規）
