+++ library/strategy_flux.py
@@ -1,4 +1,4 @@
-import os
+﻿import os
 import glob
 from typing import Any, List, Optional, Tuple, Union
 import torch
@@ -229,7 +229,7 @@ class FluxLatentsCachingStrategy(LatentsCachingStrategy):
             encode_by_vae, vae_device, vae_dtype, image_infos, flip_aug, alpha_mask, random_crop, multi_resolution=True
         )
 
-        if not train_util.HIGH_VRAM:
+        if not accelerator_setup.HIGH_VRAM:
             train_util.clean_memory_on_device(vae.device)
