+++ library/strategy_sd.py
@@ -1,11 +1,11 @@
-import glob
+﻿import glob
 import os
 from typing import Any, List, Optional, Tuple, Union
 
 import numpy as np
 import torch
 from transformers import CLIPTokenizer
-from library import train_util
+from library import accelerator_setup, train_util
 from library.strategy_base import LatentsCachingStrategy, TokenizeStrategy, TextEncodingStrategy
 from library.utils import setup_logging
 
@@ -175,5 +175,5 @@ class SdSdxlLatentsCachingStrategy(LatentsCachingStrategy):
             encode_by_vae, vae_device, vae_dtype, image_infos, flip_aug, alpha_mask, random_crop, multi_resolution=True
         )
 
-        if not train_util.HIGH_VRAM:
+        if not accelerator_setup.HIGH_VRAM:
             train_util.clean_memory_on_device(vae.device)
