+++ library/strategy_sd3.py
@@ -1,4 +1,4 @@
-import os
+﻿import os
 import glob
 import random
 from typing import Any, List, Optional, Tuple, Union
@@ -416,5 +416,5 @@ class Sd3LatentsCachingStrategy(LatentsCachingStrategy):
             encode_by_vae, vae_device, vae_dtype, image_infos, flip_aug, alpha_mask, random_crop, multi_resolution=True
         )
 
-        if not train_util.HIGH_VRAM:
+        if not accelerator_setup.HIGH_VRAM:
             train_util.clean_memory_on_device(vae.device)
